Fix block merging (#1)
Block merging should work as BFS, not DFS.  I tried to make it BFS, but
that was wrong.  It is now, I hope.

“Explanation”:

Say we have [x, [["MERGE_BLOCK"], "MERGE_BLOCK"]] and now we want to
merge a block into that.  The old way was to do BFS over every level,
but basically DFS into every level.  So we would merge the block into
the first MERGE_BLOCK, which is wrong.

We want to steadily increase the depth that we reach out into the syntax
tree instead, so we get proper BFS.